dbFix

+++ backend/database/migrations/2022_04_23_225845_add_quote_count_to_places_table.php
@@ -14,7 +14,7 @@ class AddQuoteCountToPlacesTable extends Migration
     public function up()
     {
         Schema::table('places', function (Blueprint $table) {
-            $table->bigInteger('quote_count');
+            $table->integer('quote_count');
         });
     }
